requirements.txt
minio==7.2.5
apache-airflow-providers-amazon==8.17.0
#pendulum==2.1.2
pandas==2.0.3
urllib3==1.26.7
requests==2.31.0
pyarrow==15.0.2
apache-airflow-providers-trino==5.7.0
apache-airflow-providers-common-sql==1.3.1
trino==0.328.0
apache-airflow-providers-apache-spark==4.8.0
apache-airflow-providers-ftp==3.9.0
apache-airflow-providers-sftp==4.10.0
apache-airflow-providers-ssh==3.11.0
pyspark==3.5.1
SQLAlchemy==1.4.52
paramiko==3.4.0
thrift>=0.9.2
pyhive[hive_pure_sasl]>=0.7.0
thrift>=0.9.2